Move geo types to extra
