--- python/psqlpy/geometric_types.py
@@ -1,18 +1,0 @@
-from ._internal.geometric_types import (
-    PyPoint,
-    PyBox,
-    PyPath,
-    PyLine,
-    PyLineSegment,
-    PyPolygon,
-)
-
-
-__all__ = [
-    "PyPoint",
-    "PyBox",
-    "PyPath",
-    "PyLine",
-    "PyLineSegment",
-    "PyPolygon",
-]
